Out getWinstonTransports func to winston-util.js
